add annotations

+++ MongoBase/Controllers/MongoController.cs
@@ -9,6 +9,7 @@ using System;
 using System.Web;
 using Microsoft.AspNetCore.Http;
 using MongoBase.Attributes;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace MongoBase.Controllers
 {
@@ -28,6 +29,7 @@ namespace MongoBase.Controllers
         /// <param name="query">a mongo query for syntax see : https://docs.mongodb.com/manual/tutorial/query-documents/</param>
         /// <returns>the query result</returns>
         [HttpPost("mongoquery")]
+        [SwaggerResponse(200)]
         public virtual PagedResultModel<TDocument> MongoQuery([FromBody] dynamic query,
             [FromQuery(Name = "$top")] int top = 1000,
             [FromQuery(Name = "$skip")] int skip = 0,
@@ -37,6 +39,7 @@ namespace MongoBase.Controllers
         }
 
         [HttpGet("")]
+        [SwaggerResponse(200)]
         public virtual PagedResultModel<TDocument> Get(
             [FromQuery(Name = "$top")] int top = 1000,
             [FromQuery(Name = "$skip")] int skip = 0,
@@ -73,6 +76,7 @@ namespace MongoBase.Controllers
 
 
         [HttpGet("delta")]
+        [SwaggerResponse(200)]
         public virtual IEnumerable<TDocument> delta([FromQuery] long since = 0, [FromQuery] int skip = 0, [FromQuery] int take = 200)
         {
             var query = this._repository.AsQueryable().
@@ -85,31 +89,68 @@ namespace MongoBase.Controllers
         }
         // GET api/<TestController>/5
         [HttpGet("{id}")]
-        public virtual TDocument Get(string id)
+        [SwaggerResponse(200)]
+        [SwaggerResponse(404, "NOT FOUND")]
+        public virtual ActionResult<TDocument> Get(string id)
         {
-            return this._repository.FindById(id);
+            var instance = this._repository.FindById(id);
+            if (instance == null)
+            {
+                return NotFound();
+            }
+            return instance;
         }
 
         // POST api/<TestController>
         [HttpPost]
-        public virtual string Post([FromBody] TDocument value)
+        [SwaggerResponse(200)]
+        [SwaggerResponse(409, "CONFLICT")]
+        public virtual ActionResult<string> Post([FromBody] TDocument value)
         {
-            return this._repository.InsertOne(value).Id.ToString();
+
+            var instance = this._repository.FindById(value.Id);
+            if (instance != null)
+            {
+                return Conflict("DUPLICATE KEY");
+            }
+            return this._repository.InsertOne(value).Id.ToString(); ;
         }
 
         // PUT api/<TestController>/5
         [HttpPut("{id}")]
-        public virtual void Put(string id, [FromBody] TDocument value)
+        [SwaggerResponse(204)]
+        [SwaggerResponse(409)]
+        public virtual ActionResult Put(string id, [FromBody] TDocument value)
         {
+            var instance = this._repository.FindById(value.Id);
+            if (instance == null)
+            {
+                return NotFound();
+            }
+            if (instance.ChangedAt / 1000 != value.ChangedAt / 1000)
+            {
+                return Conflict("CONCURRENT UPDATE");
+            }
             value.Id = id;
             this._repository.ReplaceOne(value);
+            return NoContent();
         }
 
-        // DELETE api/<TestController>/5
         [HttpDelete("{id}")]
-        public virtual void Delete(string id)
+        [SwaggerResponse(204)]
+        [SwaggerResponse(404)]
+        [SwaggerOperation("Delete instance by ID")]
+        public virtual ActionResult Delete(string id)
         {
+
+            var instance = this._repository.FindById(id);
+            if (instance == null)
+            {
+                return NotFound();
+            }
             this._repository.DeleteById(id);
+            return NoContent();
+
         }
